rename Poll/Proposal.ts to Poll/Poll.ts
