+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/IdeaCollection/Poll/Poll.ts
